:truck: chore(api) move certifiable badge acquisition repo
