+++ api/src/certification/shared/infrastructure/repositories/certifiable-badge-acquisition-repository.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
-import { knex } from '../../../db/knex-database-connection.js';
-import { DomainTransaction } from '../../../src/shared/domain/DomainTransaction.js';
-import { CertifiableBadgeAcquisition } from '../../../src/shared/domain/models/CertifiableBadgeAcquisition.js';
+import { knex } from '../../../../../db/knex-database-connection.js';
+import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.js';
+import { CertifiableBadgeAcquisition } from '../../../../shared/domain/models/CertifiableBadgeAcquisition.js';
 
 const BADGE_ACQUISITIONS_TABLE = 'badge-acquisitions';
